+++ avalokai/embed/chunk.py
@@ -1,6 +1,5 @@
 from langchain_core.documents.base import Document
 from langchain_text_splitters import RecursiveCharacterTextSplitter
-from transformers import AutoTokenizer
 
 from ..configs.config import Config
 
@@ -21,11 +20,3 @@ class Chunk:
         splits = self.text_splitter.split_documents(docs)
 
         return splits
-
-    # def get_chunks(self, text: str):
-    #     tokens = self.tokenizer.tokenize(text)
-    #     splits: list[list[str]] = []
-    #     for i in range(0, len(tokens), self.chunk_size - self.chunk_overlap):
-    #         splits.append(tokens[i : i + self.chunk_size])
-
-    #     return splits
