+++ avalokai/embed/hf_embed.py
@@ -2,7 +2,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 from transformers import AutoModel, AutoTokenizer
 
-from ..configs.config import Config, ModelType
+from ..configs import Config, ModelType
 from .embed import Embed
