+++ avalokai/index.py
@@ -4,9 +4,8 @@ import time
 import torch
 from tqdm import tqdm
 
-from .configs.config import Config
-from .embed import get_embedder
-from .embed.chunk import Chunk
+from .configs import Config
+from .embed import Chunk, get_embedder
 from .sink.tasks import insert_embeddings
 from .sink.vectordb import ChromaVectorDB
 from .source import (
@@ -27,18 +26,6 @@ class Indexer:
         self.embedder = get_embedder(self.config, self.device)
         self.db = ChromaVectorDB(self.config.embedding_size, dbname, create=True)
         self.dbname = dbname
-
-    # def index_single_document(self, data: RawData):
-    #     document = data.get_langchain_document()
-    #     splits = self.chunker.get_chunks([document])
-    #     contents = [split.page_content for split in splits]
-    #     metadatas = [split.metadata for split in splits]
-    #     embeddings = self.embedder.embed_multiple_documents(contents)
-
-    #     vectors: list[VectorDBData] = VectorDBData.get_data(
-    #         embeddings, metadatas, [data.id] * len(splits)
-    #     )
-    #     self.db.insert_multiple(vectors)
 
     def _index(self, dataset: BaseDataset):
         dataloader = get_data_loader(dataset, self.config.batch_size, 4)
@@ -67,8 +54,4 @@ class Indexer:
 
     def index_s3_data(self, url: str, region: str):
         dataset = get_s3_dataset(url, region, self.chunker)
-        for sample in dataset:
-            import pdb
-
-            pdb.set_trace()
         self._index(dataset)
